\name{case_1_1}
\alias{case_1_1}
\docType{data}
\title{Rickey Henderson's Batting Statistics}
\description{
Season to season batting statistics for Rickey Henderson.
}
\usage{
case_1_1
}
\format{
  A data frame.
  \describe{
  \item{Year}{Season}
  \item{Age}{Age in that season}
  \item{Tm}{Team abbreviation}
  \item{G}{Games played}
  \item{AB}{At-bats}
  \item{R}{Runs scored}
  \item{H}{Hits}
  \item{X2B}{Doubles}
  \item{X3B}{Triples}
  \item{HR}{Home runs}
  \item{RBI}{Runs batted in}
  \item{SB}{Stolen bases}
  \item{CS}{Caught stealing}
  \item{BB}{Base on balls}
  \item{SO}{Strikeouts}
  \item{BA}{Batting average}
  \item{OBP}{On-base percentage}
  \item{SLG}{Slugging percentage}
  \item{OPS}{OPS measure}
}
}
\source{Lahman database}
\keyword{datasets}
