\name{case_2_1}
\alias{case_2_1}
\docType{data}
\title{Team Batting Statistics}
\description{
Batting statistics for 30 Major League Baseball teams in 2014.
}
\usage{
case_2_1
}
\format{
  A data frame.
  \describe{
  \item{Team}{Team name}
  \item{G}{Games}
  \item{AB}{At-bats}
  \item{PA}{Plate appearances}
  \item{H}{Hits}
  \item{X1B}{Singles}
  \item{X2B}{Doubles}
  \item{X3B}{Triples}
  \item{HR}{Home runs}
  \item{R}{Runs scored}
  \item{RBI}{Runs batted in}
  \item{BB}{Base on balls}
  \item{IBB}{Intentional walks}
  \item{SO}{Strikeouts}
  \item{HBP}{Hit by pitch}
  \item{SF}{Sacrifice flies}
  \item{SH}{Sacrifice hits}
  \item{GDP}{Ground into double plays}
  \item{SB}{Stolen bases}
  \item{CS}{Caught stealing}
  \item{AVG}{Batting average}
}
}
\source{Lahman database}
\keyword{datasets}
