\name{case_2_2}
\alias{case_2_2}
\docType{data}
\title{Derek Jeter's batting statistics}
\description{
Batting statistics for Derek Jeter for all of the seasons of his career.
}
\usage{
case_2_2
}
\format{
  A data frame.
  \describe{
  \item{Year}{Team name}
  \item{Age}{Age}
  \item{G}{Games}
  \item{PA}{Plate appearances}
  \item{AB}{At-bats}
  \item{R}{Runs}
  \item{H}{Hits}
  \item{X2B}{Doubles}
  \item{X3B}{Triples}
  \item{HR}{Home runs}
  \item{RBI}{Runs batted in}
  \item{BB}{Base on balls}
  \item{SO}{Strikeouts}
  \item{BA}{Batting average}
  \item{OBP}{On-base percentage}
  \item{SLG}{Slugging percentage}
  \item{OPS}{OPS measure}
}
}
\source{Lahman database}
\keyword{datasets}
