\name{case_2_3}
\alias{case_2_3}
\docType{data}
\title{Randy Johnson's statistics}
\description{
Pitching statistics for Randy Johnson for all of the seasons of his career.
}
\usage{
case_2_3
}
\format{
  A data frame.
  \describe{
  \item{yearID}{Season}
  \item{W}{Games won}
  \item{L}{Games lost}
  \item{G}{Games pitched}
  \item{IP}{Innings pitched}
  \item{H}{Hits}
  \item{R}{Runs}
  \item{SO}{Strikeouts}
  \item{BB}{Base on balls}
  \item{ERA}{Earned Run Average}
}
}
\source{Lahman database}
\keyword{datasets}
