\name{case_2_4}
\alias{case_2_4}
\docType{data}
\title{Baseball attendance data}
\description{
Mean home attendance for all Major League Baseball teams in 2014.
}
\usage{
case_2_4
}
\format{
  A data frame.
  \describe{
  \item{HomeTeam}{Abbreviation of home team}
  \item{Mean}{Mean attendance}
}
}
\source{Lahman database}
\keyword{datasets}
