\name{case_2_5}
\alias{case_2_5}
\docType{data}
\title{Sacrifice bunt data}
\description{
Count of successful sacrifice hits for all Major League teams in 2013.
}
\usage{
case_2_5
}
\format{
  A data frame.
  \describe{
  \item{BAT_TEAM}{Abbreviation of team name}
  \item{SH}{Number of sacrifice hits}
  \item{League}{League (NL or AL)}
}
}
\source{Lahman database}
\keyword{datasets}
