\name{case_3_1}
\alias{case_3_1}
\docType{data}
\title{Hitting data for Albert Pujols and Manny Rameriz}
\description{
Hitting data for Albert Pujols and Manny Rameriz for all the seasons of their career.
}
\usage{
case_3_1
}
\format{
  A data frame.
  \describe{
  \item{Player}{Last name of player}
  \item{yearID}{Season}
  \item{AB}{At-bats}
  \item{BB}{Base on balls}
  \item{HBP}{Hit by pitch}
  \item{H}{Hits}
  \item{X2B}{Doubles}
  \item{X3B}{Triples}
  \item{HR}{Home runs}
  \item{SF}{Sacrifice flies}
  \item{Age}{Age}
  \item{SLG}{Slugging percentage}
  \item{OBP}{On-base percentage}
  \item{OPS}{OPS statistic}
}
}
\source{Lahman database}
\keyword{datasets}
