\name{case_3_2}
\alias{case_3_2}
\docType{data}
\title{Pitching data for Robin Roberts and Whitey Ford}
\description{
Pitching data for Robin Roberts and Whitey Ford for all the seasons of their career.
}
\usage{
case_3_2
}
\format{
  A data frame.
  \describe{
  \item{Year}{Season}
  \item{Player}{Last name of player}
  \item{Age}{Age}
  \item{Tm}{Team abbreviation}
  \item{W}{Wins}
  \item{L}{Losses}
  \item{G}{Games}
  \item{IP}{Innings pitched}
  \item{H}{Hits}
  \item{ER}{Earned runs}
  \item{HR}{Home runs}
  \item{BB}{Base on balls}
  \item{SO}{Strikeouts}
  \item{ERA}{Earned run average}
  \item{X.lgERA}{League ERA}
  \item{X.ERA.}{Adjusted ERA}
}
}
\source{Lahman database and Baseball Reference}
\keyword{datasets}
