\name{case_3_3}
\alias{case_3_3}
\docType{data}
\title{Team home run counts for four seasons}
\description{
Team home run counts for the 1927, 1961, 1998, and 2001 seasons.
}
\usage{
case_3_3
}
\format{
  A data frame.
  \describe{
  \item{YEAR}{Season}
  \item{HR}{Home runs}
  \item{G}{Games played}
  \item{RATE}{Home run rate}
}
}
\source{Lahman database}
\keyword{datasets}
