\name{case_3_4}
\alias{case_3_4}
\docType{data}
\title{Player slugging percentages}
\description{
Slugging percentages of all players with at least 400 at-bats in the 2014 season.
}
\usage{
case_3_4
}
\format{
  A data frame.
  \describe{
  \item{playerID}{Season}
  \item{SLG}{Slugging percentage}
}
}
\source{Lahman database}
\keyword{datasets}
