\name{case_3_5}
\alias{case_3_5}
\docType{data}
\title{Player batting averages for four seasons}
\description{
Batting averages for players with at least 400 at-bats for the 1941, 1977, 1980, and 1994 seasons.
}
\usage{
case_3_5
}
\format{
  A data frame.
  \describe{
  \item{Lastname}{Player's last name}
  \item{Firstname}{Player's first name}
  \item{Year}{Season}
  \item{AB}{At-bats}
  \item{H}{Hits}
  \item{X2B}{Doubles}
  \item{X3B}{Triples}
  \item{HR}{Home runs}
}
}
\source{Lahman database}
\keyword{datasets}
