\name{case_4_1b}
\alias{case_4_1b}
\docType{data}
\title{Home run and triple rates for historical teams}
\description{
Team names, years, home runs per game, and triples per game for 30 historical baseball teams.
}
\usage{
case_4_1b
}
\format{
  A data frame.
  \describe{
  \item{teamID}{Team abbreviation}
  \item{yearID}{Season}
  \item{HomeRun.Rate}{Number of home runs per game}
  \item{Triples.Rate}{Number of triples per game}
}
}
\source{Lahman Database}
\keyword{datasets}
