\name{case_4_2}
\alias{case_4_2}
\docType{data}
\title{Team batting statistics}
\description{
Batting statistics for all teams in the 2014 season.
}
\usage{
case_4_2
}
\format{
  A data frame.
  \describe{
  \item{Tm}{Team abbreviation}
  \item{NBat}{Number of players used in game}
  \item{BatAge}{Average age of batters}
  \item{R.G}{Average runs per game}
  \item{G}{Game}
  \item{PA}{Plate appearances}
  \item{AB}{At-bats}
  \item{R}{Runs scored}
  \item{H}{Hits}
  \item{X2B}{Doubles}
  \item{X3B}{Triples}
  \item{HR}{Home runs}
  \item{RBI}{Runs batted in}
  \item{SB}{Stolen bases}
  \item{CS}{Caught stealing}
  \item{BB}{Walks}
  \item{SO}{Strikeouts}
  \item{BA}{Batting average}
  \item{OBP}{On-base percentage}
  \item{SLG}{Slugging percentage}
  \item{OPS}{OPS}
  \item{OPS.}{Adjusted OPS}
  \item{TB}{Total bases}
  \item{GDP}{Grounded into double plays}
  \item{HBP}{Hit by pitch}
  \item{SH}{Sacrifice hits}
  \item{SF}{Sacrifice flies}
  \item{IBB}{Intentional walks}
  \item{LOB}{Total left on base}
}
}
\source{Baseball Reference}
\keyword{datasets}
