\name{case_4_3}
\alias{case_4_3}
\docType{data}
\title{Team Offensive Statistics}
\description{
Offensive statistics for the 2014 American League teams.
}
\usage{
case_4_3
}
\format{
  A data frame.
  \describe{
  \item{teamID}{Team abbreviation}
  \item{R}{Runs scored}
  \item{G}{Games played}
  \item{R.G}{Runs per game}
  \item{AVG}{Batting average}
  \item{OBP}{On-base percentage}
  \item{SLG}{Slugging percentage}
  \item{OPS}{OPS measure}
  \item{RC}{Runs created}
  \item{TA}{Total average}
  \item{BRA}{Batter's run average}
}
}
\source{Lahman Database}
\keyword{datasets}
