\name{case_4_5}
\alias{case_4_5}
\docType{data}
\title{Team records, runs, and runs against}
\description{
Team statistics for the Major League teams in the 2014 season.
}
\usage{
case_4_5
}
\format{
  A data frame.
  \describe{
  \item{teamID}{Team abbreviation}
  \item{W}{Number of wins}
  \item{L}{Number of losses}
  \item{R}{Runs scored}
  \item{RA}{Runs against}
}
}
\source{Lahman database}
\keyword{datasets}
