\name{case_4_6}
\alias{case_4_6}
\docType{data}
\title{Slugging percentages for two seasons}
\description{
2013 and 2014 slugging percentages for 20 randomly selected players.
}
\usage{
case_4_6
}
\format{
  A data frame.
  \describe{
  \item{Name}{Player name}
  \item{SLG.2013}{Slugging percentage in 2013 season}
  \item{SLG.2014}{Slugging percentage in 2014 season}
  \item{Improvement}{Change in slugging percentage}
}
}
\source{Lahman database}
\keyword{datasets}
