\name{case_5_1}
\alias{case_5_1}
\docType{data}
\title{Home run log data}
\description{
Plate appearances and home runs for Chris Davis for each game played during the 2013 baseball season.
}
\usage{
case_5_1
}
\format{
  A data frame.
  \describe{
  \item{Game}{Game number}
  \item{PA}{Number of plate appearances}
  \item{HR}{Number of home runs}
}
}
\source{Retrosheet}
\keyword{datasets}
