\name{case_6_1}
\alias{case_6_1}
\docType{data}
\title{Game to game hitting data}
\description{
2014 game to game batting results of Melky Cabrera including the number of at-bats and the number of hits.
}
\usage{
case_6_1
}
\format{
  A data frame.
  \describe{
  \item{GAME_ID}{Game id}
  \item{Date}{Date of game}
  \item{AB}{At-bats}
  \item{H}{Hits}
}
}
\source{Retrosheet}
\keyword{datasets}
