\name{case_7_5}
\alias{case_7_5}
\docType{data}
\title{Gywnn and Boggs hitting data}
\description{
Career batting statistics for Tony Gwynn and Wade Boggs.
}
\usage{
case_7_5
}
\format{
  A data frame.
  \describe{
  \item{PLAYER}{Last name of player}
  \item{AGE}{Age}
  \item{AB}{At-bats}
  \item{H}{Hits}
  \item{AVG}{Batting average}
}
}
\source{Retrosheet}
\keyword{datasets}
