\name{case_8_1}
\alias{case_8_1}
\docType{data}
\title{Situational data for Mike Trout}
\description{
Situational hitting data for Mike Trout for the 2014 season.
}
\usage{
case_8_1
}
\format{
  A data frame.
  \describe{
  \item{Season}{Season}
  \item{Situation}{Particular situation}
  \item{G}{Games}
  \item{AB}{At-bats}
  \item{PA}{Plate appearances}
  \item{H}{Hits}
  \item{X1B}{Singles}
  \item{X2B}{Doubles}
  \item{X3B}{Triples}
  \item{HR}{Home runs}
  \item{R}{Runs}
  \item{RBI}{Runs batted in}
  \item{BB}{Walks}
  \item{IBB}{Intentional walks}
  \item{SO}{Strikeouts}
  \item{HBP}{Hit by pitch}
  \item{SF}{Sacrifice flies}
  \item{SH}{Sacrifice hits}
  \item{GDP}{Groundout in double play}
  \item{SB}{Stolen bases}
  \item{CS}{Caught stealing}
  \item{AVG}{Batting average}
}
}
\source{Fangraphs}
\keyword{datasets}
