\name{case_8_2}
\alias{case_8_2}
\docType{data}
\title{Home/away batting data for two seasons}
\description{
Home and away on-base percentages for a group of 20 hitters for the 2013 and 2014 seasons.
}
\usage{
case_8_2
}
\format{
  A data frame.
  \describe{
  \item{Name}{Player name}
  \item{Home2013}{OBP for home games in 2013 season}
  \item{Away2013}{OBP for away games in 2013 season}
  \item{Home2014}{OBP for home games in 2014 season}
  \item{Away2014}{OBP for away games in 2014 season}
}
}
\source{Lahman database}
\keyword{datasets}
