\name{case_8_3}
\alias{case_8_3}
\docType{data}
\title{Player batting statistics for 2014 season}
\description{
Batting statistics and on-base percentages of all players with at least 300 at-bats in the 2014 season.
}
\usage{
case_8_3
}
\format{
  A data frame.
  \describe{
  \item{Player}{Player name}
  \item{AB}{At-bats}
  \item{H}{Hits}
  \item{BB}{Walks}
  \item{HBP}{Hit by pitch}
  \item{SF}{Sacrifice flies}
  \item{OBP}{On-base percentage}
}
}
\source{Lahman database}
\keyword{datasets}
