\name{case_8_4}
\alias{case_8_4}
\docType{data}
\title{Game-to-game hitting data for Michael Brantley}
\description{
Game-to-game hitting data for Michael Brantley for 2014 season.
}
\usage{
case_8_4
}
\format{
  A data frame.
  \describe{
  \item{Date}{Date of game}
  \item{AB}{At-bats}
  \item{H}{Hits}
}
}
\source{Retrosheet}
\keyword{datasets}
