\name{exer_2_0}
\alias{exer_2_0}
\docType{data}
\title{Batting statistics for Rickey Henderson}
\description{
Batting statistics for Rickey Henderson for all seasons of his career.
}
\usage{
exer_2_0
}
\format{
  A data frame.
  \describe{
  \item{Year}{Season}
  \item{Age}{Age}
  \item{Tm}{Team abbreviation}
  \item{G}{Games}
  \item{AB}{At-bats}
  \item{R}{Runs}
  \item{H}{Hits}
  \item{X2B}{Doubles}
  \item{X3B}{Triples}
  \item{HR}{Home runs}
  \item{RBI}{Runs batted in}
  \item{SB}{Stolen bases}
  \item{CS}{Caught stealing}
  \item{BB}{Walks}
  \item{SO}{Strikeouts}
  \item{BA}{Batting average}
  \item{OBP}{On-base percentage}
  \item{SLG}{Slugging percentage}
  \item{OPS}{OPS}
}
}
\source{Lahman database}
\keyword{datasets}
