\name{exer_2_12}
\alias{exer_2_12}
\docType{data}
\title{Batting statistics for Bob Feller}
\description{
Pitching statistics for Bob Feller for all seasons of his career.
}
\usage{
exer_2_12
}
\format{
  A data frame.
  \describe{
  \item{YR}{Season}
  \item{TM}{Team abbreviation}
  \item{Lg}{League}
  \item{GP}{Games pitched}
  \item{W}{Number of wins}
  \item{L}{Number of losses}
  \item{PCT}{Winning percentage}
  \item{IP}{Innings pitched}
  \item{H}{Hits}
  \item{BB}{Walks}
  \item{SO}{Strikeouts}
  \item{ERA}{Earned run average}
}
}
\source{Lahman database}
\keyword{datasets}
