\name{exer_2_18}
\alias{exer_2_18}
\docType{data}
\title{Lengths of baseball games}
\description{
Length (in minutes) of a sample of 60 baseball games
in 2014.
}
\usage{
exer_2_18
}
\format{
  A data frame.
  \describe{
  \item{Game}{Game number}
  \item{Time}{Length of game}
}
}
\source{Lahman database}
\keyword{datasets}
