\name{exer_2_20}
\alias{exer_2_20}
\docType{data}
\title{Game home runs}
\description{
Number of home runs hit in a sample of 2014 baseball games.
}
\usage{
exer_2_20
}
\format{
  A data frame.
  \describe{
  \item{Game}{Game number}
  \item{Number_HR}{Number of home runs hit}
}
}
\source{Retrosheet}
\keyword{datasets}
