\name{exer_2_21}
\alias{exer_2_21}
\docType{data}
\title{Team batting statistics}
\description{
Batting statistics for the 2014 American League teams.
}
\usage{
exer_2_21
}
\format{
  A data frame.
  \describe{
  \item{Team}{Game number}
  \item{G}{Number of home runs hit}
  \item{AB}{At-bats}
  \item{PA}{Plate appearances}
  \item{H}{Hits}
  \item{X1B}{Singles}
  \item{X2B}{Doubles}
  \item{X3B}{Triples}
  \item{HR}{Home runs}
  \item{R}{Runs}
  \item{RBI}{Runs batted in}
  \item{BB}{Walks}
  \item{IBB}{Intentional walks}
  \item{SO}{Strikeouts}
  \item{HBP}{Hit by pitch}
  \item{SF}{Sacrifice flies}
  \item{SH}{Sacrifice hits}
  \item{GDP}{Grounded in double play}
  \item{SB}{Stolen base}
  \item{CS}{Caught stealing}
  \item{AVG}{Batting average}
  \item{L}{League indicator}
}
}
\source{Lahman database}
\keyword{datasets}
