\name{exer_2_23}
\alias{exer_2_23}
\docType{data}
\title{HOF Pitching Data}
\description{
Birthyear and ERA for pitchers inducted in the baseball Hall of Fame.
}
\usage{
exer_2_23
}
\format{
  A data frame.
  \describe{
  \item{Pitcher}{Pitcher name}
  \item{Birthyear}{Year of birth}
  \item{ERA}{Earned run average}
}
}
\source{Lahman database}
\keyword{datasets}
