\name{exer_2_25}
\alias{exer_2_25}
\docType{data}
\title{Salaries of the Atlanta Braves}
\description{
Salaries (in dollars) of the players on the 2014 Atlanta Braves.
}
\usage{
exer_2_25
}
\format{
  A data frame.
  \describe{
  \item{Player}{Player name}
  \item{Salary}{Salary in dollars}
}
}
\source{Baseball Reference}
\keyword{datasets}
