\name{exer_2_29}
\alias{exer_2_29}
\docType{data}
\title{Baseball attendance}
\description{
Average attendance figures for all ML teams in 2014.
}
\usage{
exer_2_29
}
\format{
  A data frame.
  \describe{
  \item{HomeTeam}{Team name abbreviation}
  \item{League}{League abbreviation}
  \item{Attendance}{Mean attendance}
}
}
\source{Lahman database}
\keyword{datasets}
