\name{exer_2_31}
\alias{exer_2_31}
\docType{data}
\title{Batting and throwing hands}
\description{
Batting side and throwing hand for forty randomly selected players that were born 1950 or later.
}
\usage{
exer_2_31
}
\format{
  A data frame.
  \describe{
  \item{Batting.Hand}{Side of batting, L, R, or B}
  \item{Throwing.Hand}{Side of pitching, L or R}
}
}
\source{Lahman database}
\keyword{datasets}
