\name{exer_2_32}
\alias{exer_2_32}
\docType{data}
\title{Fastball speeds}
\description{
Speeds of fastballs (in mph) thrown by Zack Greinke during a specific game during the 2015 season.
}
\usage{
exer_2_32
}
\format{
  A data frame.
  \describe{
  \item{Pitch}{Pitch number}
  \item{Speed}{Pitch speed in mph}
}
}
\source{PitchFX data}
\keyword{datasets}
