\name{exer_2_33}
\alias{exer_2_33}
\docType{data}
\title{Pitch types}
\description{
Pitch types of the first 40 pitches thrown by Zack Greinke during a specific game thrown during the 2015 season.
}
\usage{
exer_2_33
}
\format{
  A data frame.
  \describe{
  \item{Pitch}{Pitch number}
  \item{Pitch_Type}{Abbreviation of pitch type}
}
}
\source{PitchFX data}
\keyword{datasets}
