\name{exer_2_5}
\alias{exer_2_5}
\docType{data}
\title{Batting statistics for Babe Ruth}
\description{
Batting statistics for Babe Ruth for all seasons of his career.
}
\usage{
exer_2_5
}
\format{
  A data frame.
  \describe{
  \item{YR}{Season}
  \item{TM}{Team abbreviation}
  \item{L}{League}
  \item{G}{Games}
  \item{AB}{At-bats}
  \item{R}{Runs}
  \item{H}{Hits}
  \item{HR}{Home runs}
  \item{RBI}{Runs batted in}
  \item{AVG}{Batting average}
  \item{OBP}{On-base percentage}
  \item{SLG}{Slugging percentage}
}
}
\source{Lahman database}
\keyword{datasets}
