\name{exer_3_0}
\alias{exer_3_0}
\docType{data}
\title{Henderson and Raines Data}
\description{
On-base and slugging percentages for Rickey Henderson and Tim Raines for all seasons of their career.
}
\usage{
exer_3_0
}
\format{
  A data frame.
  \describe{
  \item{PLAYER}{Last name of player}
  \item{AGE}{Age}
  \item{OBP}{On-base percentage}
  \item{SLG}{Slugging percentage}
}
}
\source{Lahman database}
\keyword{datasets}
