\name{exer_3_10}
\alias{exer_3_10}
\docType{data}
\title{Team Home Run Counts}
\description{
Home run numbers of all Major League teams in 2014.
}
\usage{
exer_3_10
}
\format{
  A data frame.
  \describe{
  \item{teamID}{Team abbreviation}
  \item{lgID}{League abbreviation}
  \item{HR}{Home runs}
}
}
\source{Lahman database}
\keyword{datasets}
