\name{exer_3_19}
\alias{exer_3_19}
\docType{data}
\title{Suzuki Batted ball statistics}
\description{
Batted ball statistics for Ichiro Suzuki for seasons 2002 through 2015.
}
\usage{
exer_3_19
}
\format{
  A data frame.
  \describe{
  \item{Season}{Season}
  \item{Pull}{Proportion of pulled balls in play}
  \item{Cent}{Proportion of balls hit in the middle}
  \item{Oppo}{Proportion of balls hit to the opposite side}
  \item{Soft}{Proportion of soft-hit balls}
  \item{Med}{Proportion of medium-hit balls}
  \item{Hard}{Proportion of hard-hit balls}
}
}
\source{Fangraphs}
\keyword{datasets}
