\name{exer_3_7}
\alias{exer_3_7}
\docType{data}
\title{Pitching data for Greg Maddux and Tom Glavine}
\description{
Pitching data for Greg Maddux and Tom Glavine for all the seasons of their career.
}
\usage{
exer_3_7
}
\format{
  A data frame.
  \describe{
  \item{Pitcher}{Last name of player}
  \item{Year}{Season}
  \item{W}{Wins}
  \item{L}{Losses}
  \item{Win_pct}{Winning percentage}
  \item{ERA}{Earned run average}
}
}
\source{Lahman database}
\keyword{datasets}
