\name{exer_3_8}
\alias{exer_3_8}
\docType{data}
\title{Batting data for shortstops and 3rd basemen}
\description{
Slugging percentages and batting averages for regular shortstops and 3rd basemen in the 2014 season.
}
\usage{
exer_3_8
}
\format{
  A data frame.
  \describe{
  \item{LastName}{Last name of player}
  \item{FirstName}{First name of player}
 \item{Position}{Fielding position}
 \item{SLG}{Slugging percentage}
 \item{AVG}{Batting average}
}
}
\source{Lahman database}
\keyword{datasets}
