\name{exer_3_9}
\alias{exer_3_9}
\docType{data}
\title{Batting statistics of 1897 and 1997 players}
\description{
Batting statistics of all players with at least 400 at-bats
from 1897 and 1997 seasons.
}
\usage{
exer_3_9
}
\format{
  A data frame.
  \describe{
  \item{Lastname}{Last name of player}
  \item{Firstname}{First name of player}
  \item{Year}{Season}
  \item{AB}{At-bats}
  \item{H}{Hits}
  \item{X2B}{Doubles}
  \item{X3B}{Triples}
  \item{HR}{Home runs}
}
}
\source{Lahman database}
\keyword{datasets}
