\name{exer_4_0}
\alias{exer_4_0}
\docType{data}
\title{Rickey Henderson Data}
\description{
On-base and slugging percentages for Rickey Henderson for all seasons of his career.
}
\usage{
exer_4_0
}
\format{
  A data frame.
  \describe{
  \item{AGE}{Age}
  \item{OBP}{On-base percentage}
  \item{SLG}{Slugging percentage}
}
}
\source{Lahman database}
\keyword{datasets}
