\name{exer_4_1}
\alias{exer_4_1}
\docType{data}
\title{Average and runs scored}
\description{
Batting average and runs scored per game for 2014 National League teams.
}
\usage{
exer_4_1
}
\format{
  A data frame.
  \describe{
  \item{teamID}{Team abbreviation}
  \item{AVG}{Batting average}
  \item{R.G}{Runs scored per game}
}
}
\source{Lahman database}
\keyword{datasets}
