\name{exer_4_12}
\alias{exer_4_12}
\docType{data}
\title{Population and baseball attendance}
\description{
United States population, Major League baseball attendance, and ratio for the years 1900 through 2010.
}
\usage{
exer_4_12
}
\format{
  A data frame.
  \describe{
  \item{Year}{Season}
  \item{Population}{United States population (millions)}
  \item{Attendance}{Total MLB baseball attendance}
  \item{Ratio}{Ratio of attendance over population}
}
}
\source{Lahman database and almanac}
\keyword{datasets}
