\name{exer_4_13}
\alias{exer_4_13}
\docType{data}
\title{Season batting stats for two players}
\description{
Batting statistics of Jose Altuve and Nelson Cruz for the 2014 baseball season.
}
\usage{
exer_4_13
}
\format{
  A data frame.
  \describe{
  \item{Player}{Player}
  \item{AB}{At-bats}
  \item{H}{Hits}
  \item{X2B}{Doubles}
  \item{X3B}{Triples}
  \item{HR}{Home runs}
  \item{BB}{Walks}
  \item{HBP}{Hit by pitch}
  \item{SF}{Sacrifice flies}
}
}
\source{Lahman database}
\keyword{datasets}
