\name{exer_4_15}
\alias{exer_4_15}
\docType{data}
\title{Team home runs}
\description{
National League team home runs for the 2014 baseball season.
}
\usage{
exer_4_15
}
\format{
  A data frame.
  \describe{
  \item{Team}{Team abbreviation}
  \item{HR}{Home runs}
}
}
\source{Lahman database}
\keyword{datasets}
