\name{exer_4_17}
\alias{exer_4_17}
\docType{data}
\title{Batting statistics for teams}
\description{
National League batting statistics for the 2014 baseball season.
}
\usage{
exer_4_17
}
\format{
  A data frame.
  \describe{
  \item{Team}{Team abbreviation}
  \item{R.G}{Runs scored per game}
  \item{AVG}{Batting average}
  \item{SLG}{Slugging percentage}
  \item{OPS}{OPS measure}
}
}
\source{Lahman database}
\keyword{datasets}
