\name{exer_4_20}
\alias{exer_4_20}
\docType{data}
\title{Batted ball data and slugging percentages}
\description{
Proportion of hard batted balls and the slugging percentages for 14 seasons of the career of Albert Pujols.
}
\usage{
exer_4_20
}
\format{
  A data frame.
  \describe{
  \item{Season}{Season}
  \item{Hard}{Proportion of hard-hit balls}
  \item{SLG}{Slugging percentage}
}
}
\source{Lahman database}
\keyword{datasets}
