\name{exer_4_21}
\alias{exer_4_21}
\docType{data}
\title{Pitches and game duration}
\description{
Number of pitches and game duration in minutes for twenty games in the 2015 season.
}
\usage{
exer_4_21
}
\format{
  A data frame.
  \describe{
  \item{Game}{Game number}
  \item{Pitches}{Total number of pitches}
  \item{Duration}{Length of game in minutes}
}
}
\source{Lahman database}
\keyword{datasets}
