\name{exer_4_3}
\alias{exer_4_3}
\docType{data}
\title{Team Pitching Data}
\description{
Team pitching statistics for 2014 National League teams.
}
\usage{
exer_4_3
}
\format{
  A data frame.
  \describe{
  \item{teamID}{Team abbreviation}
  \item{SOR}{Number of pitching strikeouts per game}
  \item{BOR}{Number of pitching walks per game}
  \item{ERA}{Earned run average}
  \item{PCT}{Winning percentage}
  \item{HG}{Number of hits allowed per game}
  \item{HRG}{Number of home runs allowed per game}
  \item{RG}{Number of runs allowed per game}
}
}
\source{Lahman database}
\keyword{datasets}
