\name{exer_5_19}
\alias{exer_5_19}
\docType{data}
\title{Batting Statistics for All-Stars}
\description{
Career batting statistics for 18 great players.
}
\usage{
exer_5_19
}
\format{
  A data frame.
  \describe{
  \item{Name}{Player's name}
  \item{Position}{Fielding position}
  \item{League}{League abbreviation}
  \item{G}{Games}
  \item{AB}{At-bats}
  \item{R}{Runs}
  \item{H}{Hits}
  \item{X2B}{Doubles}
  \item{X3B}{Triples}
  \item{HR}{Home runs}
  \item{SO}{Strikeouts}
  \item{BB}{Walks}
  \item{AVG}{Batting average}
  \item{OBP}{On-base percentage}
  \item{SLG}{Slugging percentage}
}
}
\source{Retrosheet}
\keyword{datasets}
