\name{exer_5_4}
\alias{exer_5_4}
\docType{data}
\title{Game by game strikeouts}
\description{
Number of strikeouts and innings pitched for all games Madison Bumgarner started in the 2014 season.
}
\usage{
exer_5_4
}
\format{
  A data frame.
  \describe{
  \item{Game}{Game number}
  \item{IP}{Innings pitched}
  \item{H}{Hits}
  \item{R}{Runs allowed}
  \item{ER}{Earned runs}
  \item{BB}{Walks}
  \item{SO}{Strikeouts}
}
}
\source{Lahman database}
\keyword{datasets}
