\name{exer_5_8}
\alias{exer_5_8}
\docType{data}
\title{Runners and Runs in Simulated Baseball}
\description{
Two-way count table of the number of runners on base and the runs scored in half-innings from 1000 simulated games of Big League Baseball.
}
\usage{
exer_5_8
}
\format{
  A data frame.
  \describe{
  \item{Runners}{Number of runners in inning}
  \item{Runs}{Runs scored in inning}
  \item{Count}{Number of half-innings observed}
}
}
\source{Retrosheet}
\keyword{datasets}
