\name{exer_5_9}
\alias{exer_5_9}
\docType{data}
\title{Runners and Runs in Real Baseball}
\description{
Two-way count table of the number of runners on base and the runs scored in half-innings from ten consecutive Phillies games in 1999.
}
\usage{
exer_5_9
}
\format{
  A data frame.
  \describe{
  \item{Runners}{Number of runners in inning}
  \item{Runs}{Runs scored in inning}
  \item{Count}{Number of half-innings observed}
}
}
\source{Retrosheet}
\keyword{datasets}
