\name{exer_7_1}
\alias{exer_7_1}
\docType{data}
\title{Batting Averages for Two Seasons}
\description{
Batting averages for 12 randomly selected playes in the 2013 and 2014 seasons.
}
\usage{
exer_7_1
}
\format{
  A data frame.
  \describe{
  \item{Player}{Player's name}
  \item{AVG.2013}{Batting average in 2013 season}
  \item{AVG.2014}{Batting average in 2014 season}
}
}
\source{Lahman package}
\keyword{datasets}
