\name{exer_8_1}
\alias{exer_8_1}
\docType{data}
\title{Home and Away Batting Averages}
\description{
Home and away batting averages for 14 randomly selected players from the 2013 season.
}
\usage{
exer_8_1
}
\format{
  A data frame.
  \describe{
  \item{Name}{Player's name}
  \item{AB_Home}{At-bats at home}
  \item{H_Home}{Hits at home}
  \item{AVG_Home}{Batting average at home}
  \item{AB_Away}{At-bats at away games}
  \item{H_Away}{Hits at away games}
  \item{AVG_Away}{Batting average at away games}
}
}
\source{Lahman package}
\keyword{datasets}
