\name{exer_8_2}
\alias{exer_8_2}
\docType{data}
\title{Pre and Post All-Star Batting Averages}
\description{
Home and away batting averages for 15 randomly selected players before and after the All-Star Break from the 2013 season.
}
\usage{
exer_8_2
}
\format{
  A data frame.
  \describe{
  \item{Name}{Player's name}
  \item{AB_Pre}{At-bats before the All-Star break}
  \item{H_Pre}{Hits before the All-Star break}
  \item{AVG_Pre}{Batting average before the All-Star break}
  \item{AB_Post}{At-bats after the All-Star break}
  \item{H_Post}{Hits after the All-Star break}
  \item{AVG_Post}{Batting average after the All-Star break}
}
}
\source{Lahman package}
\keyword{datasets}
