\name{exer_8_3}
\alias{exer_8_3}
\docType{data}
\title{The Initial Pitch and Batting Averages}
\description{
Batting statistics for 15 players with an initial ball (1-0 count) and an initial strike (0-1 count) for 2013 season.
}
\usage{
exer_8_3
}
\format{
  A data frame.
  \describe{
  \item{Name}{Player's name}
  \item{AB_Ball}{At-bats with an initial ball}
  \item{H_Ball}{Hits with an initial ball}
  \item{AVG_Ball}{Batting average with an initial ball}
  \item{AB_Strike}{At-bats with an initial strike}
  \item{H_Strike}{Hits with an initial strike}
  \item{AVG_Strike}{Batting average with an initial strike}
}
}
\source{Lahman package}
\keyword{datasets}
