\name{exer_8_4}
\alias{exer_8_4}
\docType{data}
\title{Batting Averages on Even and Odd Numbered Days}
\description{
Simulated batting statistics for 15 players on even-numbered and odd-numbered days.
}
\usage{
exer_8_4
}
\format{
  A data frame.
  \describe{
  \item{NAME}{Player's name}
  \item{EVEN_AB}{At-bats on "even" days}
  \item{EVEN_H}{Hits on "even" days}
  \item{EVEN_AVG}{Batting average on "even" days}
  \item{ODD_AB}{At-bats on "odd" days}
  \item{ODD_H}{Hits on "odd" days}
  \item{ODD_AVG}{Batting average on "odd" days}
}
}
\source{Lahman package}
\keyword{datasets}
