\name{exer_8_5}
\alias{exer_8_5}
\docType{data}
\title{Batting Averages with there is a Situational Bias}
\description{
Simulated batting statistics for 15 players using special and usual baseballs when there is a situational bias.
}
\usage{
exer_8_5
}
\format{
  A data frame.
  \describe{
  \item{NAME}{Player's name}
  \item{SP_AB}{At-bats with special ball}
  \item{SP_H}{Hits with special ball}
  \item{SP_AVG}{Batting average with special ball}
  \item{U_AB}{At-bats with usual ball}
  \item{U_H}{Hits with usual ball}
  \item{U_AVG}{Batting average with usual ball}
}
}
\source{Lahman package}
\keyword{datasets}
