\name{exer_8_6}
\alias{exer_8_6}
\docType{data}
\title{Batting Averages on Even and Odd Numbered Days}
\description{
Simulated batting statistics for 15 players using special and usual baseballs when there is situational ability.
}
\usage{
exer_8_6
}
\format{
  A data frame.
  \describe{
  \item{U_AB}{At-bats with usual ball}
  \item{U_H}{Hits with usual ball}
  \item{U_AVG}{Batting average with usual ball}
  \item{NAME}{Player's name}
  \item{S_AB}{At-bats with special ball}
  \item{S_H}{Hits with special ball}
  \item{S_AVG}{Batting average with special ball}
}
}
\source{Lahman package}
\keyword{datasets}
