\name{exer_8_7}
\alias{exer_8_7}
\docType{data}
\title{Home and Away Batting Statistics}
\description{
Home and away batting averages for 14 randomly selected players from the 2014 season.
}
\usage{
exer_8_7
}
\format{
  A data frame.
  \describe{
  \item{Name}{Player's name}
  \item{AB_Home}{At-bats at home games}
  \item{H_Home}{Hits at home games}
  \item{AVG_Home}{Batting average at home games}
  \item{AB_Away}{At-bats at away games}
  \item{H_Away}{Hits at away games}
  \item{AVG_Away}{Batting average at away games}
}
}
\source{Lahman package}
\keyword{datasets}
