\name{exer_8_8}
\alias{exer_8_8}
\docType{data}
\title{Situational Pitching Data}
\description{
Situational statistics for 20 pitchers for the three-year period 1997-1999.
}
\usage{
exer_8_8
}
\format{
  A data frame.
  \describe{
  \item{Pitcher}{Player's name}
  \item{Throws}{Throwing side of pitcher}
  \item{Left_AVG}{Batting average against left-handed batters}
  \item{Right_AVG}{Batting average against right-handed batters}
  \item{X1.15_AVG}{Batting average on pitches 1 through 15}
  \item{X46.60_AVG}{Batting average on pitches 46 through 60}
  \item{X91.105_AVG}{Batting average on pitches 91 through 105}
}
}
\source{Fangraphs}
\keyword{datasets}
